WIP: get rid of temporary lib/yargs-types
